Set drawer text colors based on dining hall events
Green for Farm Fridays/Healthy Mondays, blue for college nights, and
light gray for closed.

Also improved logic for which meal tab to show on selecting new dining
hall - may want to change this some more so that it shows the same tab
instead of simply not showing breakfast when the brunch message is
there.